Add files via upload
Changes:

Selling flow: 
Created an input boolean so that the checks in others flow if electricity is beeing sold, does not depend on a figure. 
It now depends on a boolean on or off.

zero export flow is beeing simplified.
It is easier to understand.

When importprice is negative onlmy start charging during this time and charge battery on the negative
This needs to be tested still!